Add a generic encode method
